+++ src/xefis/support/protocols/nmea/parser.cc
@@ -132,7 +132,7 @@ Parser::verify_sentence (std::string const& sentence)
 		// Verify checksum:
 		uint8_t expected_checksum = 0;
 		for (auto c = sentence.begin() + 1; c != sentence.end() - 3; ++c)
-			expected_checksum ^= *c;
+			expected_checksum ^= static_cast<uint8_t> (*c);
 
 		if (expected_checksum != parsed_checksum)
 			throw nmea::InvalidChecksum (expected_checksum, parsed_checksum);
